Use .pc files as sort order hints (#27)
vcpkg-rs discovers libraries to link using the manifest files from
vcpkg. The manifest files do not indicate any ordering of library files.

While ports are handled correctly already in dependency order, multiple
libraries defined by a single Port are otherwise not sorted. This can
cause problems when there is an explicit ordering that should be obeyed.

This is observed when building rust-sdl2 using the 'use-vcpkg' flag, as
this transitively includes the brotli port, which defines 3 libraries:
libbrotlicommon, libbrotlidec and libbrotlienc. These libraries expect
to be linked with libbrotlicommon last, as otherwise rustc's invocation
of link will fail (as --as-needed is passed to the linker).

See also: https://github.com/microsoft/vcpkg/issues/16991
See also: https://github.com/Rust-SDL2/rust-sdl2/pull/1085

This commit works around the issue by teaching vcpkg-rs the bare minimum
required to parse (optional) .pc files if they are available in the
vcpkg package, and uses them to infer the correct ordering that should
be used.

Closes #26