Remove Cargo.lock
Keeping a Cargo.lock file around doesn't make much sense for this workspace,
nor for vcpkg as an intermediate library.

The lockfile may have been important at some point when trying to keep
the vcpkg-rs crate compatible with early downversions of rustc, but even
then the tests against rust 1.12 do not rely on contents of this
lockfile.